Разделены функции виджетов по отдельным модулям
